+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v1.0.12 (TBA)
+
+* `PowPersistentSession.Plug.Base` now accepts `:persistent_session_ttl` which will pass the TTL to the cache backend and used for the max age of the sesion cookie in `PowPersistentSession.Plug.Cookie`
+* Deprecated `:persistent_session_cookie_max_age` configuration setting
+
 ## v1.0.11 (2019-06-13)
 
 * Fixed bug in router filters with Phoenix 1.4.7